Removed unnecessary require_once() calls, reorganized tests
