+++ tests/CollectionTest.php
@@ -1,9 +1,29 @@
 <?php
 
-require_once("../danielgsims/Collection.php");
-require_once("testClass.php");
-
 use danielgsims\Collection;
+
+class TestClassA{
+  private $v;
+  public function __construct($v){
+    $this->setValue($v);
+  }
+
+  public function getValue(){
+    return $this->v;
+  }
+
+  public function setValue($v){
+    $this->v = $v;
+  }
+}
+
+class TestClassB{
+
+}
+
+class TestClassExtendsA extends TestClassA{
+
+}
 
 class ControllerTest extends PHPUnit_Framework_TestCase {
